backend/requirements.txt
fastapi
uvicorn
websockets
openai
python-dotenv
typing-extensions